refactor(skill-index): move SkillIndexEntry to hermes impl, rename adaptor class, shorten tool description
Move SkillIndexEntry from the source-agnostic adaptor interface into
hermes-skill-index.ts since it's specific to the Hermes JSON shape.
Rename HermesSkillIndexAdaptor to HermesSkillIndex to match the
FileAdaptor/LocalFiles naming convention (impl drops the Adaptor
suffix). Shorten the searchSkills tool description to one sentence.
