+++ apps/dashboard/src/server/usecases/adaptors/skill-index.ts
@@ -1,16 +1,3 @@
-export interface SkillIndexEntry {
-  name: string;
-  description: string;
-  source: string;
-  identifier: string;
-  trust_level: string;
-  repo?: string;
-  path?: string;
-  tags: string[];
-  extra?: { provider?: string; [k: string]: unknown };
-  resolved_github_id?: string;
-}
-
 export interface SkillSearchResult {
   name: string;
   identifier: string;
